pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cursorconnect"
version = "0.1.0"
description = "Python SDK for CursorConnect"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
    { name = "Joshua Reed" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests>=2.28",
]

[project.optional-dependencies]
test = [
    "python-dotenv",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["cursorconnect*"]

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (requires live Cursor API credentials; deselect with -m 'not slow')",
]
filterwarnings = [
    "ignore::DeprecationWarning:cursorconnect",
]
